+++ spec/client_spec.rb
@@ -57,29 +57,29 @@ describe Service::Client do
     end
 
     it "calls the right url with the right method after adding it" do
-      must_send_request(:post, 'http://example.com/authors/', {name: 'Peter Lustig'}, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}"}) do
+      must_send_request(:post, 'http://example.com/authors/', {name: 'Peter Lustig'}, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}", 'Content-Type' => 'application/json'}) do
         @client.post(@client.urls.author, TOKEN, name: 'Peter Lustig')
       end
 
-      must_send_request(:get, 'http://example.com/authors/123', nil, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}"}) do
+      must_send_request(:get, 'http://example.com/authors/123', nil, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}", 'Content-Type' => 'application/json'}) do
         @client.get(@client.urls.author(123), TOKEN)
       end
 
-      must_send_request(:post, 'http://example.com/authors/123/books/456', {name: 'Ronald Review', comment: 'This book is the bomb!'}, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}"}) do
+      must_send_request(:post, 'http://example.com/authors/123/books/456', {name: 'Ronald Review', comment: 'This book is the bomb!'}, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}", 'Content-Type' => 'application/json'}) do
         @client.post(@client.urls.review(author_id: 123, book_id: 456), TOKEN, name: 'Ronald Review', comment: 'This book is the bomb!')
       end
 
-      must_send_request(:post, 'http://example.com/authors/123/books/456', {name: 'Ronald Review', comment: 'This book is the bomb!'}, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}"}) do
+      must_send_request(:post, 'http://example.com/authors/123/books/456', {name: 'Ronald Review', comment: 'This book is the bomb!'}, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}", 'Content-Type' => 'application/json'}) do
         @client.post(@client.urls.review(123, 456), TOKEN, name: 'Ronald Review', comment: 'This book is the bomb!')
       end
     end
 
     it "uses query parameters instead of JSON bodies for GET requests" do
-      must_send_request(:get, 'http://example.com/authors/123?some=arguments&are=cool%26not%3Dbody&array%5B%5D=12&array%5B%5D=gla&hash%5Bblub%5D=123&hash%5Bbla%5D=moep', nil, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}"}) do
+      must_send_request(:get, 'http://example.com/authors/123?some=arguments&are=cool%26not%3Dbody&array%5B%5D=12&array%5B%5D=gla&hash%5Bblub%5D=123&hash%5Bbla%5D=moep', nil, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}", 'Content-Type' => 'application/json'}) do
         @client.get(@client.urls.author(123), TOKEN, {some: 'arguments', are: "cool&not=body", array: ["12", 'gla'], hash: {"blub" => "123", "bla" => "moep"}})
       end
 
-      must_send_request(:get, 'http://example.com/authors/123/another-fixed-part?blub=1&some=arguments&are=cool%26not%3Dbody', nil, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}"}) do
+      must_send_request(:get, 'http://example.com/authors/123/another-fixed-part?blub=1&some=arguments&are=cool%26not%3Dbody', nil, headers: {'AUTHORIZATION' => "Bearer #{TOKEN}", 'Content-Type' => 'application/json'}) do
         @client.get(@client.urls.author_with_query(123), TOKEN, {some: 'arguments', are: "cool&not=body"})
       end
     end